change to tmp directory instead of files, force pnm
Unpaper requires input and output files to be specified with a number
format (e.g. input%d.pnm) so separation of pages in double-page scans will
work.  Therefore, we change from individual temporary files to a temporary
directory to have more control over the file names.  Also, unpaper requires
the input files to be in the PNM format, so we force this.
* scanner.el (scanner-scan-document): force PNM, use tmp directories